correct parameter type

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -8,7 +8,7 @@ import random
 async def wait_random(max_delay: int = 10) -> float:
     """asynchronous coroutine
     param:
-        @max_delay: float
+        @max_delay: int
     return:
         float
     """
